requirements.txt
minio
polygon-api-client
pandas
pyarrow
s3fs
